Refactor of card component
